+++ bin/faucet/src/main.rs
@@ -48,8 +48,9 @@ const ENV_REMOTE_TX_PROVER_URL: &str = "MIDEN_FAUCET_REMOTE_TX_PROVER_URL";
 const ENV_POW_SECRET: &str = "MIDEN_FAUCET_POW_SECRET";
 const ENV_POW_CHALLENGE_LIFETIME: &str = "MIDEN_FAUCET_POW_CHALLENGE_LIFETIME";
 const ENV_POW_CLEANUP_INTERVAL: &str = "MIDEN_FAUCET_POW_CLEANUP_INTERVAL";
-const ENV_POW_GROWTH_RATE: &str = "MIDEN_FAUCET_POW_GROWTH_RATE";
+const ENV_POW_CHALLENGES_PER_DIFFICULTY: &str = "MIDEN_FAUCET_CHALLENGES_PER_DIFFICULTY";
 const ENV_POW_BASELINE: &str = "MIDEN_FAUCET_POW_BASELINE";
+const ENV_POW_BASE_DIFFICULTY_AMOUNT: &str = "MIDEN_FAUCET_POW_BASE_DIFFICULTY_AMOUNT";
 const ENV_API_KEYS: &str = "MIDEN_FAUCET_API_KEYS";
 const ENV_ENABLE_OTEL: &str = "MIDEN_FAUCET_ENABLE_OTEL";
 const ENV_STORE: &str = "MIDEN_FAUCET_STORE";
@@ -88,7 +89,7 @@ pub enum Command {
         #[arg(long = "account", value_name = "FILE", env = ENV_ACCOUNT_PATH)]
         faucet_account_path: PathBuf,
 
-        /// The maximum amount of assets base units that can be dispersed on each request.
+        /// The maximum amount of assets' base units that can be dispersed on each request.
         #[arg(long = "max-claimable-amount", value_name = "U64", env = ENV_MAX_CLAIMABLE_AMOUNT, default_value = "1000000000")]
         max_claimable_amount: u64,
 
@@ -112,10 +113,13 @@ pub enum Command {
         #[arg(long = "pow-challenge-lifetime", value_name = "DURATION", env = ENV_POW_CHALLENGE_LIFETIME, default_value = "30s", value_parser = humantime::parse_duration)]
         pow_challenge_lifetime: Duration,
 
-        /// A measure of how quickly the `PoW` difficult grows with the number of requests. When
-        /// set to 1, the difficulty will roughly double when the number of requests doubles.
-        #[arg(long = "pow-growth-rate", value_name = "NON_ZERO_USIZE", env = ENV_POW_GROWTH_RATE, default_value = "1")]
-        pow_growth_rate: NonZeroUsize,
+        /// A measure of how quickly the `PoW` difficulty increases with the number of requests.
+        /// Represents how many requests are needed to increase the load difficulty by 1. As
+        /// reference, the target for the challenges is computes as:
+        /// * `load_difficulty = (num_active_challenges / challenges_per_difficulty) + 1`
+        /// * `max_target / (load_difficulty * request_complexity)`
+        #[arg(long = "pow-challenges-per-difficulty", value_name = "NON_ZERO_USIZE", env = ENV_POW_CHALLENGES_PER_DIFFICULTY, default_value = "10")]
+        pow_challenges_per_difficulty: NonZeroUsize,
 
         /// The interval at which the `PoW` challenge cache is cleaned up.
         #[arg(long = "pow-cleanup-interval", value_name = "DURATION", env = ENV_POW_CLEANUP_INTERVAL, default_value = "2s", value_parser = humantime::parse_duration)]
@@ -123,10 +127,20 @@ pub enum Command {
 
         /// The baseline for the `PoW` challenges. This sets the `PoW` difficulty (in bits) that a
         /// a challenge will have when there are no requests against the faucet. It must be between
-        /// 0 and 32.
-        #[arg(value_parser = clap::value_parser!(u8).range(0..=32))]
-        #[arg(long = "pow-baseline", value_name = "U8", env = ENV_POW_BASELINE, default_value = "12")]
+        /// 0 and 64.
+        #[arg(value_parser = clap::value_parser!(u8).range(0..=64))]
+        #[arg(long = "pow-baseline", value_name = "U8", env = ENV_POW_BASELINE, default_value = "16")]
         pow_baseline: u8,
+
+        /// The base difficulty amount for the `PoW` challenges. This sets the requested amount at
+        /// which the difficulty of the challenges will start to increase.
+        ///
+        /// The target of the challenges is divided by a scaling factor based on the requested
+        /// amount:
+        /// * `request_complexity = (amount / base_difficulty_amount) + 1`
+        /// * `max_target / (load_difficulty * request_complexity)`
+        #[arg(long = "pow-base-difficulty-amount", value_name = "U64", env = ENV_POW_BASE_DIFFICULTY_AMOUNT, default_value = "100000000")]
+        pow_base_difficulty_amount: u64,
 
         /// Comma-separated list of API keys.
         #[arg(long = "api-keys", value_name = "STRING", env = ENV_API_KEYS, num_args = 1.., value_delimiter = ',')]
@@ -217,8 +231,9 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
             pow_secret,
             pow_challenge_lifetime,
             pow_cleanup_interval,
-            pow_growth_rate,
+            pow_challenges_per_difficulty,
             pow_baseline,
+            pow_base_difficulty_amount,
             api_keys,
             open_telemetry: _,
             store_path,
@@ -259,7 +274,7 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
             let rate_limiter_config = PoWRateLimiterConfig {
                 challenge_lifetime: pow_challenge_lifetime,
                 cleanup_interval: pow_cleanup_interval,
-                growth_rate: pow_growth_rate,
+                challenges_per_difficulty: pow_challenges_per_difficulty,
                 baseline: pow_baseline,
             };
 
@@ -269,6 +284,7 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
                 max_supply,
                 faucet.issuance(),
                 max_claimable_amount,
+                pow_base_difficulty_amount,
                 tx_mint_requests,
                 pow_secret.as_str(),
                 rate_limiter_config,
@@ -503,8 +519,9 @@ mod test {
                         pow_secret: "test".to_string(),
                         pow_challenge_lifetime: Duration::from_secs(30),
                         pow_cleanup_interval: Duration::from_secs(1),
-                        pow_growth_rate: NonZeroUsize::new(1).unwrap(),
+                        pow_challenges_per_difficulty: NonZeroUsize::new(1).unwrap(),
                         pow_baseline: 12,
+                        pow_base_difficulty_amount: 100_000,
                         faucet_account_path: faucet_account_path.clone(),
                         remote_tx_prover_url: None,
                         open_telemetry: false,